# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Doc API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/authors": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Список",
                "operationId": "be68aa64058e29169345590ef50213d8",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "name"
                                                    },
                                                    "bio": {
                                                        "type": "string",
                                                        "example": "bio"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2024-11-24T14:38:49.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2024-11-24T14:38:49.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Authors"
                ],
                "summary": "Создание",
                "operationId": "bcab5786e00765bc5a28d96b6ec4f48b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/AuthorStoreRequestSchema"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "name"
                                                },
                                                "bio": {
                                                    "type": "string",
                                                    "example": "bio"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-11-24T14:38:49.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-11-24T14:38:49.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/authors/{author}": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Просмотр",
                "operationId": "13c2a81fa6f28d2b4c6f11b375c01258",
                "parameters": [
                    {
                        "name": "author",
                        "in": "path",
                        "description": "ID автора",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "name"
                                                },
                                                "bio": {
                                                    "type": "string",
                                                    "example": "bio"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-11-24T14:38:49.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-11-24T14:38:49.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Authors"
                ],
                "summary": "Удалить",
                "operationId": "dd7ad00391cba1f67c97ca7cc2aca01c",
                "parameters": [
                    {
                        "name": "destroy",
                        "in": "path",
                        "description": "ID автора",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Author deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Authors"
                ],
                "summary": "Обновление",
                "operationId": "b96ef33e375a5e6260763a428a819b59",
                "parameters": [
                    {
                        "name": "patсh",
                        "in": "path",
                        "description": "ID автора",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/AuthorUpdateRequestSchema"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "name"
                                                },
                                                "bio": {
                                                    "type": "string",
                                                    "example": "bio"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-11-24T14:38:49.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2024-11-24T14:38:49.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/authors/books/{author}": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Книги автора",
                "operationId": "a6b2b21b59aa38f821141b207fd9acd2",
                "parameters": [
                    {
                        "name": "author",
                        "in": "path",
                        "description": "ID автора",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Some title"
                                            },
                                            "author_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "published_at": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2012-11-25"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2024-11-24T14:38:49.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2024-11-24T14:38:49.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AuthorStoreRequestSchema": {
                "required": [
                    "name",
                    "bio"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "name"
                    },
                    "bio": {
                        "type": "string",
                        "example": "bio"
                    }
                },
                "type": "object"
            },
            "AuthorUpdateRequestSchema": {
                "required": [
                    "name",
                    "bio"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "name"
                    },
                    "bio": {
                        "type": "string",
                        "example": "bio"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}